move shared stats types
