+++ entrypoints/popup/views/stats/__tests__/ReviewHistoryChart.test.tsx
@@ -6,8 +6,8 @@ import { render, screen } from '@testing-library/react';
 import { Rating } from 'ts-fsrs';
 import { describe, expect, it, vi } from 'vitest';
 import { statsQueryKeys } from '@/hooks/queries/stats';
-import type { DailyStats } from '@/services/stats';
 import { sendMessage } from '@/shared/messages';
+import type { DailyStats } from '@/shared/stats';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { ReviewHistoryChart } from '../ReviewHistoryChart';
